move getContrastTextColor to util/colors
