+++ src/components/Chart/SerendieChartProps.tsx
@@ -4,6 +4,7 @@ import { PieSvgProps } from "@nivo/pie";
 import { token } from "../../../styled-system/tokens";
 import serendieTokens from "@serendie/design-token";
 import { SerendieChartTheme } from "./SerendieChartTheme";
+import { getContrastTextColor } from "../../utils/colors";
 
 export type ChartColorCategory =
   | "primary"
@@ -72,31 +73,6 @@ export const getChartColor = (
   index: number = 0
 ): string => {
   return getChartColors(category)[index] || getChartColors(category)[0];
-};
-
-// 明度を計算してコントラストの高い文字色を返す関数
-export const getContrastTextColor = (backgroundColor: string): string => {
-  // varから始まる場合はcomputedStyleを使用
-  if (backgroundColor.startsWith("var(") && backgroundColor.endsWith(")")) {
-    // varを外す
-    const varName = backgroundColor.replace("var(", "").replace(")", "");
-    const computedStyle = getComputedStyle(document.documentElement);
-    backgroundColor = computedStyle.getPropertyValue(varName);
-  }
-
-  // 背景色をRGBに変換
-  const hex = backgroundColor.replace("#", "");
-  const r = parseInt(hex.substr(0, 2), 16);
-  const g = parseInt(hex.substr(2, 2), 16);
-  const b = parseInt(hex.substr(4, 2), 16);
-
-  // 明度を計算 (YIQアルゴリズム)
-  const yiq = (r * 299 + g * 587 + b * 114) / 1000;
-
-  // 明度が128以上なら黒、それ以下なら白
-  return yiq >= 128
-    ? token("colors.sd.system.color.chart.component.onMarkLabel")
-    : token("colors.sd.system.color.chart.component.inverseOnMarkLabel");
 };
 
 const spacingTokens = serendieTokens.sd.system.dimension.spacing;
